beans + logic fixes

+++ src/main/java/by/epamtc/enrollmentsystem/service/Facilitym2mUserInfoService.java
@@ -5,5 +5,5 @@ import by.epamtc.enrollmentsystem.exception.ServiceException;
 import java.util.List;
 
 public interface Facilitym2mUserInfoService {
-    List<String> getUserFacilitiesNames(int userId) throws ServiceException;
+    List<String> getUserFacilitiesNames(long userId) throws ServiceException;
 }
